redirect to homepage when not logged in
Somehow google started listing the app homepage vs the
openstax.org/kinetic page
