+++ backend/app/controllers/static_controller.rb
@@ -1,11 +1,16 @@
 # frozen_string_literal: true
 
 class StaticController < ActionController::Base
+
   def catchall
-    if Rails.env.production?
+    redirect_to local_dev_path_for_request and return unless Rails.env.production?
+
+    if has_auth_cookie?
       render file: 'public/index.html'
     else
-      redirect_to Rails.application.secrets.frontend_url + request.fullpath
+      # see other is treated as signal that target should be canonical
+      # https://developers.google.com/search/docs/crawling-indexing/301-redirects
+      redirect_to Rails.application.secrets.homepage_url, status: :see_other
     end
   end
 
@@ -18,5 +23,13 @@ class StaticController < ActionController::Base
         render text: 'bad request, content not found', status: 404
       end
     end
+  end
+
+  protected
+
+  include CookieAuthentication
+
+  def local_dev_path_for_request
+    Rails.application.secrets.frontend_url + request.fullpath
   end
 end
